# oidc2-e2e-server/src/go/model_error_status.go
/*
 * Endpoint for establishing and maintaining an encrypted client-server session
 *
 * This endpoint establishes an end-to-end encrypted client-server session after authenticating the client with its Id Certification Token (Client_ICT). After the successful exchange of DH parameters, the client can communicate with the server, upload, download or delete files on the server with all communication being encrypted on application layer.
 *
 * API version: 0.1
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger

// Information about ocurred error.
type ErrorStatus struct {
	// Status Code
	Code int32 `json:"code"`
	// Status Text
	Status string `json:"status"`
	// More detailed description
	Description string `json:"description,omitempty"`
}